# apps/backend/docs/openapi.yml
openapi: 3.0.3
info:
  title: Gestiabloc API
  version: 1.0.0
  description: |
    Authentication and file upload endpoints for Gestiabloc.
servers:
  - url: http://localhost:4000
    description: Local backend
...
paths:
...
  /api/account-settings/{tenantId}/{section}:
    get:
      tags: [Account Settings]
      summary: List account-settings resources by section
      description: |
        For `section=users`, returns tenant members with pagination.
        For `section=account`, returns tenant profile information.
        For other sections, returns `501 NOT_IMPLEMENTED` until implemented.
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantIdPathParam"
        - $ref: "#/components/parameters/AccountSettingsSectionPathParam"
        - $ref: "#/components/parameters/AccountSettingsPageQueryParam"
        - $ref: "#/components/parameters/AccountSettingsPageSizeQueryParam"
      responses:
        "200":
          description: Returned when `section=users` or `section=account`
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/AccountSettingsUsersListResponse"
                  - $ref: "#/components/schemas/AccountSettingsTenantInfoResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not a tenant admin for this tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "501":
          description: Endpoint scaffolded but not implemented yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountSettingsNotImplementedCollectionResponse"
    patch:
      tags: [Account Settings]
      summary: Update tenant information
      description: Only valid when `section=account`.
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantIdPathParam"
        - $ref: "#/components/parameters/AccountSettingsSectionPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountSettingsTenantInfoUpdateRequest"
      responses:
        "200":
          description: Tenant profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountSettingsTenantInfoResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not a tenant admin for this tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Tenant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Account Settings]
      summary: Create account-settings resource in section
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantIdPathParam"
        - $ref: "#/components/parameters/AccountSettingsSectionPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountSettingsMutationRequest"
      responses:
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not a tenant admin for this tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "501":
          description: Endpoint scaffolded but not implemented yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountSettingsNotImplementedCollectionResponse"
  /api/account-settings/{tenantId}/{section}/{recordId}:
    put:
      tags: [Account Settings]
      summary: Replace account-settings record in section
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantIdPathParam"
        - $ref: "#/components/parameters/AccountSettingsSectionPathParam"
        - $ref: "#/components/parameters/AccountSettingsRecordIdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountSettingsMutationRequest"
      responses:
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not a tenant admin for this tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "501":
          description: Endpoint scaffolded but not implemented yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountSettingsNotImplementedRecordResponse"
    patch:
      tags: [Account Settings]
      summary: Update account-settings record in section
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantIdPathParam"
        - $ref: "#/components/parameters/AccountSettingsSectionPathParam"
        - $ref: "#/components/parameters/AccountSettingsRecordIdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountSettingsMutationRequest"
      responses:
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not a tenant admin for this tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "501":
          description: Endpoint scaffolded but not implemented yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountSettingsNotImplementedRecordResponse"
    delete:
      tags: [Account Settings]
      summary: Delete account-settings record in section
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantIdPathParam"
        - $ref: "#/components/parameters/AccountSettingsSectionPathParam"
        - $ref: "#/components/parameters/AccountSettingsRecordIdPathParam"
      responses:
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not a tenant admin for this tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "501":
          description: Endpoint scaffolded but not implemented yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountSettingsNotImplementedRecordResponse"
  /api/account-settings/{tenantId}/users/{userId}/security-level:
    patch:
      tags: [Account Settings]
      summary: Update member security level
      description: Tenant-admin only. Tenant admins are fixed at `MAX`.
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantIdPathParam"
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberSecurityLevelRequest"
      responses:
        "200":
          description: Security level updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateMemberSecurityLevelResponse"
        "400":
          description: Invalid request or invalid level for tenant admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not a tenant admin for this tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Member not found for tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sid
  parameters:
    TenantIdPathParam:
      in: path
      name: tenantId
      required: true
      schema:
        type: string
    AccountSettingsSectionPathParam:
      in: path
      name: section
      required: true
      schema:
        type: string
        enum:
          - users
          - account
          - services
          - professionals
          - follow-ups
          - status-config
          - features
          - subscription
          - custom-fields
    AccountSettingsRecordIdPathParam:
      in: path
      name: recordId
      required: true
      schema:
        type: string
    AccountSettingsPageQueryParam:
      in: query
      name: page
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    AccountSettingsPageSizeQueryParam:
      in: query
      name: pageSize
      required: false
      schema:
        type: integer
        enum: [10, 25]
        default: 10
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "INVALID_CREDENTIALS"
        details:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
...
    AccountSettingsMutationRequest:
      type: object
      properties:
        tenantId:
          type: string
          description: Optional fallback tenant id. Path tenantId is authoritative.
      additionalProperties: true
    AccountSettingsNotImplementedCollectionResponse:
      type: object
      properties:
        error:
          type: string
          example: NOT_IMPLEMENTED
        tenantId:
          type: string
        section:
          type: string
        method:
          type: string
          example: POST
        scope:
          type: string
          example: collection
    AccountSettingsNotImplementedRecordResponse:
      type: object
      properties:
        error:
          type: string
          example: NOT_IMPLEMENTED
        tenantId:
          type: string
        section:
          type: string
        recordId:
          type: string
        method:
          type: string
          example: PATCH
        scope:
          type: string
          example: record
    AccountSettingsUsersListResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/AccountSettingsUserItem"
        pagination:
          $ref: "#/components/schemas/AccountSettingsPagination"
        timezone:
          type: string
          nullable: true
          description: Tenant timezone (IANA). If null, clients should use system timezone.
    AccountSettingsUserItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        avatar:
          type: string
          nullable: true
        emailVerified:
          type: boolean
        isOnline:
          type: boolean
        sessionCreatedAt:
          type: string
          format: date-time
          nullable: true
        role:
          type: string
          enum: [TENANT_ADMIN, TENANT_USER]
        accountStatus:
          type: string
          enum: [ACTIVE, DISABLED]
        securityLevel:
          type: string
          enum: [LOW, MEDIUM, MAX]
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
    AccountSettingsPagination:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
          enum: [10, 25]
        total:
          type: integer
        totalPages:
          type: integer
    AccountSettingsTenantInfoUpdateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        addressLine1:
          type: string
          nullable: true
        addressLine2:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        postalCode:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        timezone:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
    AccountSettingsTenantInfoResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        tenant:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            email:
              type: string
              nullable: true
            phone:
              type: string
              nullable: true
            addressLine1:
              type: string
              nullable: true
            addressLine2:
              type: string
              nullable: true
            city:
              type: string
              nullable: true
            state:
              type: string
              nullable: true
            postalCode:
              type: string
              nullable: true
            country:
              type: string
              nullable: true
            timezone:
              type: string
              nullable: true
            website:
              type: string
              nullable: true
            emailVerified:
              type: boolean
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    UpdateMemberSecurityLevelRequest:
      type: object
      required: [securityLevel]
      properties:
        securityLevel:
          type: string
          enum: [LOW, MEDIUM, MAX]
    UpdateMemberSecurityLevelResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        membership:
          type: object
          properties:
            userId:
              type: string
            tenantId:
              type: string
            role:
              type: string
              enum: [TENANT_ADMIN, TENANT_USER]
            status:
              type: string
              enum: [ACTIVE, DISABLED]
            securityLevel:
              type: string
              enum: [LOW, MEDIUM, MAX]